+++ src/RA_Integration.vcxproj.filters
@@ -405,6 +405,9 @@
     <ClCompile Include="ui\viewmodels\ProgressTrackerViewModel.cpp">
       <Filter>UI\ViewModels</Filter>
     </ClCompile>
+    <ClCompile Include="services\AchievementRuntimeExports.cpp">
+      <Filter>Services</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="RA_Resource.h">
@@ -967,6 +970,9 @@
     </ClInclude>
     <ClInclude Include="ui\viewmodels\ProgressTrackerViewModel.hh">
       <Filter>UI\ViewModels</Filter>
+    </ClInclude>
+    <ClInclude Include="services\AchievementRuntimeExports.hh">
+      <Filter>Services</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
