+++ src/services/AchievementRuntimeExports.hh
@@ -1,0 +1,7 @@
+#ifndef RA_SERVICES_ACHIEVEMENT_RUNTIME_EXPORTS_HH
+#define RA_SERVICES_ACHIEVEMENT_RUNTIME_EXPORTS_HH
+#pragma once
+
+void SyncClientExternalRAIntegrationMenuItem(int nMenuItemId);
+
+#endif // !RA_SERVICES_ACHIEVEMENT_RUNTIME_EXPORTS_HH
